requirements.txt
Django
djangorestframework
requests
pgeocode
numpy
pandas